+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## fugit 0.9.1  not yet released
 
+* Implement Fugit .parse, .parse_at and .parse_cron
+
+
 ## fugit 0.9.0  released 2017-01-03
 
 * Initial release